Added CVector, tests, bla.

+++ ReadingData.h
@@ -3,5 +3,5 @@
 struct ReadingData
 {
 	int reading_time;
-	double reading_value;
+	float reading_value;
 };
